feat: useDetailMap CustomHooks Implement
