+++ src/components/features/CommentForm.jsx
@@ -68,6 +68,16 @@ const CommentForm = ({ id }) => {
   const handleSubmitForm = (e) => {
     e.preventDefault();
 
+    if (!comment) {
+      Swal.fire({
+        icon: 'info',
+        title: '댓글을 입력해 해주세요.',
+        text: '로그인 이후에 리뷰를 남길 수 있습니다.'
+      });
+
+      return;
+    }
+
     if (isLogin) {
       handleCommentSubmit(comment);
       setComment('');
